// TaxiDetailsClient/src/app/api/model/driver.ts
/**
 * TaxiDetails.WebApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Driver { 
    id?: number;
    name?: string | null;
    surname?: string | null;
    patronymic?: string | null;
    passport?: string | null;
    address?: string | null;
    phone?: string | null;
}