// TaxiDetailsClient/src/app/api/model/driverDto.ts
/**
 * TaxiDetails.WebApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Data transfer object for driver information.
 */
export interface DriverDto { 
    /**
     * name driver
     */
    name: string;
    /**
     * surname driver
     */
    surname: string;
    /**
     * patronymic driver
     */
    patronymic: string;
    /**
     * passport information
     */
    passport: string;
    /**
     * addres
     */
    address: string;
    /**
     * number of phone
     */
    phone: string;
}